REFACTOR: Port discrete double theories to new VDC-based trait.
