+++ packages/catlog/src/dbl/mod.rs
@@ -8,4 +8,3 @@ pub mod model;
 pub mod model_diagram;
 pub mod model_morphism;
 pub mod theory;
-pub mod theory_next;
